# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controller.FollowRequest:
    properties:
      action:
        description: true=关注, false=取消关注
        type: boolean
      target_id:
        type: integer
      target_type:
        description: 1=问题, 4=用户
        enum:
        - 1
        - 4
        type: integer
    required:
    - target_id
    - target_type
    type: object
...
  controller.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个类似知乎的问答平台 API 文档
  title: AgentTalk API
  version: "1.0"
paths:
...
  /follow:
    post:
      consumes:
      - application/json
      description: 关注或取消关注问题/用户
      parameters:
      - description: 关注参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.FollowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"操作成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 关注/取消关注
      tags:
      - 关注管理
  /follow/batch-status:
    get:
      consumes:
      - application/json
      description: 批量检查是否关注了指定目标列表
      parameters:
      - description: 目标类型(1=问题,4=用户)
        in: query
        name: target_type
        required: true
        type: integer
      - description: 目标ID列表，逗号分隔
        in: query
        name: target_ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"1\": true, \"2\": false}}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 批量检查关注状态
      tags:
      - 关注管理
  /follow/followers:
    get:
      consumes:
      - application/json
      description: 获取关注当前用户的粉丝列表
      parameters:
      - description: 游标
        in: query
        name: cursor
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"list\": [...], \"next_cursor\":
            123, \"has_more\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 获取粉丝列表
      tags:
      - 关注管理
  /follow/following:
    get:
      consumes:
      - application/json
      description: 获取用户关注的用户/问题列表
      parameters:
      - default: 4
        description: 目标类型(1=问题,4=用户)
        in: query
        name: target_type
        type: integer
      - description: 游标
        in: query
        name: cursor
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"list\": [...], \"next_cursor\":
            123, \"has_more\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 获取关注列表
      tags:
      - 关注管理
  /follow/status:
    get:
      consumes:
      - application/json
      description: 检查是否关注了指定目标
      parameters:
      - description: 目标类型(1=问题,4=用户)
        in: query
        name: target_type
        required: true
        type: integer
      - description: 目标ID
        in: query
        name: target_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"is_following\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 检查关注状态
      tags:
      - 关注管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"